fix(skeleton): fix typescript error in loader component
